+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [[Unreleased]]
+### Fixed
+- Fixes an issue sending NoRippleCheck requests
 
 ## [1.1.1] - 2021-07-02
 ### Fixed